+++ tests/unit/engines/query/test_vector_query_engine.py
@@ -54,5 +54,5 @@ class TestVectorQueryEngine:
     def test_load_artifacts(self, engine):
         engine.upsert_text_artifacts(artifacts=[TextArtifact("foobar1"), TextArtifact("foobar2")], namespace="test")
 
-        assert len(engine.load_artifacts("doesntexist")) == 0
-        assert len(engine.load_artifacts("test")) == 2
+        assert len(engine.vector_store_driver.load_artifacts("doesntexist")) == 0
+        assert len(engine.vector_store_driver.load_artifacts("test")) == 2